fix 🐛: (cli) Fix fresh installs bug close #64
