+++ packages/cli/package.json
@@ -25,29 +25,30 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "axios": "0.27.2",
+    "axios": "1.1.2",
     "bin-links": "3.0.3",
-    "chalk": "5.0.1",
+    "chalk": "5.1.0",
     "execa": "6.1.0",
     "glob": "8.0.3",
     "markdown-table": "3.0.2",
     "npm-pick-manifest": "7.0.2",
     "ora": "6.1.2",
-    "pacote": "13.6.2",
+    "pacote": "14.0.0",
     "prompts": "2.4.2",
-    "read-package-json-fast": "2.0.3",
-    "semver": "7.3.7",
+    "semver": "7.3.8",
+    "spinnies": "0.5.1",
     "tar": "6.1.11"
   },
   "devDependencies": {
     "@swc/cli": "0.1.57",
-    "@swc/core": "1.2.242",
+    "@swc/core": "1.3.6",
     "@types/glob": "7.2.0",
     "@types/mv": "2.1.2",
-    "@types/tar": "6.1.3",
     "@types/pacote": "11.1.5",
     "@types/prompts": "2.0.14",
     "@types/semver": "7.3.12",
+    "@types/spinnies": "^0.5.0",
+    "@types/tar": "6.1.3",
     "chokidar": "3.5.3"
   }
 }
