+++ packages/cli/src/utils/checkLock.ts
@@ -1,0 +1,27 @@
+import path from "node:path";
+import { getDeps } from "./getDeps.js";
+import readPackage from "./readPackage.js";
+import semver from "semver";
+
+export default function checkLock(lock: any) {
+  const pkg = readPackage(path.join(process.cwd(), "package.json"));
+  const deps = getDeps(pkg);
+
+  for (const dep of deps) {
+    if (!lock[dep.name])
+      throw new Error(`Package ${dep.name} not found in lock file`);
+    let results = [];
+    for (const version of Object.keys(lock[dep.name])) {
+      // If no version satisfies the version of the pkg, throw error
+      if (semver.satisfies(version, dep.version)) {
+        results.push(version);
+      }
+    }
+    if (results.length === 0)
+      throw new Error(
+        `No version of ${dep.name} satisfies the version ${dep.version} in package.json`
+      );
+  }
+
+  return true;
+}
