+++ packages/cli/src/utils/extract.ts
@@ -1,6 +1,8 @@
-import { existsSync, writeFileSync } from "node:fs";
+import tar from "tar";
+import axios from "axios";
+import { createWriteStream, mkdirSync, existsSync, writeFileSync } from "fs";
+import os from "os";
 import path from "path";
-import pacote from "pacote";
 import {
   __DOWNLOADED,
   __DOWNLOADING,
@@ -8,20 +10,24 @@ import {
   __SKIPPED,
   downloadFile,
 } from "../commands/install.js";
+import ora from "ora";
+import chalk from "chalk";
 
-export async function extract(
-  cacheFolder: string,
-  tarball: string
-): Promise<any> {
-  // Read .fnpm file to know if it's fully installed
-  const fnpmFile = path.join(cacheFolder, downloadFile);
-  const fnpmFileExists = existsSync(fnpmFile);
+// Get system temp directory
+const tmpDir = os.tmpdir();
 
+const cacheBasePath = path.join(tmpDir, "ultra_tmp");
+
+export async function ultraExtract(target: string, tarball: string) {
   if (!tarball) {
     throw new Error("No tarball provided");
   }
 
-  if (fnpmFileExists) {
+  // Read .ultra file to know if it's fully installed
+  const ultraFile = path.join(target, downloadFile);
+  const ultraFileExists = existsSync(ultraFile);
+
+  if (ultraFileExists || __DOWNLOADING.includes(tarball)) {
     return {
       res: "skipped",
     };
@@ -29,13 +35,51 @@ export async function extract(
 
   __DOWNLOADING.push(tarball);
 
-  // Extract tarball
-  await pacote.extract(tarball, cacheFolder).catch((err: any) => {
-    throw new Error(`Error extracting ${tarball} - ${err.message}`);
-  });
+  let file = path.join(
+    cacheBasePath,
+    // @ts-ignore-next-line
+    tarball.split("/").pop()
+  );
 
-  // Create .fnpm file
-  writeFileSync(fnpmFile, "{}");
+  if (!existsSync(cacheBasePath)) {
+    mkdirSync(cacheBasePath);
+  }
+
+  if (!existsSync(file)) {
+    const writer = createWriteStream(file);
+    const response = await axios({
+      url: tarball,
+      method: "GET",
+      responseType: "stream",
+    });
+
+    response.data.pipe(writer);
+
+    await new Promise((resolve, reject) => {
+      writer.on("finish", resolve);
+      writer.on("error", reject);
+    });
+  }
+
+  // Extract "package" directory from tarball to "target" directory
+  mkdirSync(target, { recursive: true });
+
+  await tar
+    .extract({
+      file,
+      cwd: target,
+      strip: 1,
+    })
+    .catch((err) => {
+      ora(
+        chalk.red(
+          `Error extracting ${file} to ${target}: ${err.message || err}`
+        )
+      ).fail();
+    });
+
+  // Create .ultra file
+  writeFileSync(ultraFile, "{}");
 
   __DOWNLOADING.splice(__DOWNLOADING.indexOf(tarball), 1);
