+++ packages/cli/types/index.d.ts
@@ -1,5 +1,4 @@
 declare module "npm-pick-manifest";
-declare module "read-package-json-fast";
 declare module "@npmcli/arborist";
 declare module "bin-links" {
   /**
